+++ R/utils_api.R
@@ -67,6 +67,10 @@ read_secure_json <- function(url, token = NULL, timeout = 60, config = list()) {
 #'   list of simulation parameters returned by the function
 #'   \code{\link{setSimulationPathAPI}}
 #' @param timeout \code{numeric} API timeout (seconds). Default to 60.
+#' 
+#' @return
+#' Object of class `simOptions`, list of options used to read the data contained in the last
+#' simulation read by \code{\link{setTimeoutAPI}}.
 #'
 #' @export
 #'