+++ tests/testthat/test-readClusterDesc.R
@@ -69,7 +69,7 @@ test_that("test read cluster st-storage v860", {
   path_study_test <- grep(pattern = "test_case_study_v870", 
                           x = studyPathSV8, value = TRUE)
   opts_study_test <- setSimulationPath(path_study_test, simulation = "input")
-  
+
   # function setSimulationPath() provide areas names with st-storage clusters
   areas_st <- opts_study_test$areasWithSTClusters