+++ NEWS.md
@@ -12,7 +12,7 @@ NEW FEATURES:
 It contains a table with group dimensions of time series for binding constraints.
 * `readAntares()` new parameter **clustersST** to read (output simulation) short-term clusters
 * New function `readAntaresSTClusters()`
-
+* `fread_antares()` shiny compatible with a conditional processing of the error messages
 
 BREAKING CHANGES :
 
@@ -25,6 +25,7 @@ BUGFIXES :
 * `readInputThermal()` return data from file data.txt with `thermalData` parameter
 * `setSimulationPath()` has also the parameter **areasWithSTClusters** in 'output' mode
 * `setSimulationPathAPI()`: control the existence of the output folder **links** or **areas** before reading the data (upgrade Antares Web)
+* `readClusterDesc()` / `readClusterResDesc()` / `readClusterSTDesc()` return a data.table in API mode
 
 
 # antaresRead 2.7.0